Implemented basic connect() and notify().
Using helper class M_EventFunctionBase as base class for templated class M_EventFunction. Every call() to a pointer to M_EventFunctionBase will resolve to the correct M_EventFunction via virtual functions.
The EventNotifier::m_functions is a map that binds typeid(event).hash_code() to another map, that binds the hash of the event (using std::hash) to a vector of M_EventFunctionBase*.