Added Handler to handle connected functions.
It will disconnect that function on destruction, or when disconnect() is called. keep() removes the ownership from the Handler, so that disconnect() and the destructor do nothing (to only be used when the connected functions do not own resources that are destructed before the main EventNotifier is).

Every connect*() function returns a Handler.